fix(computegraph): the group offset is widened after the addition, not before
A member's index in the flattened operand lists is a group's first slot plus an
offset within it, and both are ints, so casting the sum widened a number that had
already been computed at int width. Harmless at the counts a grouped batch holds
and wrong in principle, which is what clang-tidy says about it; each operand is
widened on its own now.

+++ libs/Einsums/ComputeGraph/src/TensorExpr.cpp
@@ -685,7 +685,7 @@ expected<ExprStatement, RaiseFailure> raise_grouped(Graph const &graph, Node con
         // shape a member runs at is its group's.
         for (auto const &group : desc->groups) {
             for (int slot = 0; slot < group.count; ++slot) {
-                auto const member = static_cast<std::size_t>(group.first + slot);
+                auto const member = static_cast<std::size_t>(group.first) + static_cast<std::size_t>(slot);
                 note_extent(family, li, member, static_cast<std::size_t>(group.m));
                 note_extent(family, lj, member, static_cast<std::size_t>(group.n));
                 note_extent(family, lk, member, static_cast<std::size_t>(group.k));
